+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] - 2017-11-26
+
 - Restructured project 
 - Removed modernizr fallback for cssanimations
 - Fixed csslint/eslint warnings
@@ -24,7 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Forked original code from codrops
 
-[Unreleased]: https://github.com/rejas/multilevelmenu/compare/1.0.1...HEAD
+[Unreleased]: https://github.com/rejas/multilevelmenu/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/rejas/multilevelmenu/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/rejas/multilevelmenu/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/rejas/multilevelmenu/compare/0.9.0...1.0.0
 [0.9.0]: https://github.com/rejas/multilevelmenu/compare/e9508983c033fd61cf25ce45f1b722d1aa2ed212...0.9.0